A coordinate is sixteen bits, in all eighteen routines that take one
Every routine taking an x and a y took them as uint32_t. They are the
positions the original keeps in registers, and cga_at already says what
that width is: sixteen bits, because plot_pixel's column arrives in CX
and reaches 319 while the rest are bytes widened into it. One rule for
all of them beats two.

Eighteen definitions, seventeen declarations, io_mouse_warp in sdl_io.c,
and entity_anim's draw callback, whose type has to match the four
routines passed to it.

The masks on arguments stay and are not made redundant by this: the
original works those out in byte registers, so `(y + 0x0b) & 0xff` still
has to wrap before it widens.

80 bytes of .text, so it is checked: draw_sprite_20x6, probe_cell_at,
ball_draw, shot_xor, pixel_xor, xor_sprite_16x7, sprite_shift_draw,
xor_sprite_20x16, xor_sprite_16xn, entity_sparkle, brick_11,
brick_11_after, cell_hole_draw, ending_plot, ending_particles_tick,
menu_particles_tick and particle_init all agree with the original, and
2,000 frames on level 11, 1,200 on level 50 and 1,200 through the bonus
are identical throughout.

+++ reconstruct/src/sdl_io.c
@@ -314,7 +314,7 @@ void io_pin_mouse(uint32_t x, uint32_t buttons)
 }
 
 
-void io_mouse_warp(uint32_t x, uint32_t y)
+void io_mouse_warp(uint16_t x, uint16_t y)
 {
     /* A warp is what the next read returns, in lockstep as much as here: the
      * play loop centres the pointer at 1ac2:1925 before the serve, and a port
